ai-parenting-assistant/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM相关
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5

# 向量数据库
chromadb==0.4.22

# 文本处理
tiktoken==0.5.2
pypdf==4.0.1
python-docx==1.1.0

# 前端
streamlit==1.30.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0
